services/confluence-importer/requirements.txt
atlassian-python-api
markitdown
requests
pytest
